+++ bdtsim/cli/environment_info.py
@@ -18,6 +18,7 @@
 import argparse
 from typing import Dict
 
+from bdtsim.account import AccountFile
 from bdtsim.environment import EnvironmentManager
 from .command_manager import SubCommand
 
@@ -28,16 +29,21 @@ class EnvironmentInfoSubCommand(SubCommand):
     def __init__(self, parser: argparse.ArgumentParser) -> None:
         super(EnvironmentInfoSubCommand, self).__init__(parser)
         parser.add_argument('environment', choices=EnvironmentManager.environments.keys())
+        parser.add_argument('--account-file', help='Specify account configuration file to be used')
         parser.add_argument('-e', '--environment-parameter', nargs=2, action='append', dest='environment_parameters',
                             default=[], metavar=('KEY', 'VALUE'), help='Pass additional parameters to the environment')
 
     def __call__(self, args: argparse.Namespace) -> int:
+        account_file = AccountFile(path=args.account_file)
         environment_parameters: Dict[str, str] = {}
         for key, value in args.environment_parameters:
             environment_parameters[key.replace('-', '_')] = value
 
         environment = EnvironmentManager.instantiate(
-            args.environment,
+            name=args.environment,
+            operator=account_file.operator,
+            seller=account_file.seller,
+            buyer=account_file.buyer,
             **environment_parameters
         )
 
@@ -49,4 +55,15 @@ class EnvironmentInfoSubCommand(SubCommand):
             print('Sync Status: syncing (%d/%d)' % (syncing.currentBlock, syncing.highestBlock))
 
         print('Chain ID: %s' % str(environment.web3.eth.chainId))
+
+        for account in account_file.operator, account_file.seller, account_file.buyer:
+            balance = environment.web3.eth.getBalance(account.wallet_address)
+            print('Account balance % 8s (%s): % 20i (~%.2f Eth, %i transactions)' % (
+                account.name,
+                account.wallet_address,
+                balance,
+                (balance / 1000000000000000000),
+                environment.web3.eth.getTransactionCount(account.wallet_address)
+            ))
+
         return 0